Move smart field service into services package
